@@ -22,16 +22,21 @@
         }
     ],
     "require": {
-        "php": ">=5.6.4",
-        "guzzlehttp/guzzle": "~6.0"
+        "php": "^7.2",
+        "guzzlehttp/guzzle": "^6.3.1|^7.0"
     },
     "require-dev": {
-        "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "~5.0"
+        "mockery/mockery": "^1.3.1",
+        "phpunit/phpunit": "^8.4|^9.0"
     },
     "autoload": {
         "psr-4": {
-            "Themsaid\\Forge\\": "src/"
+            "Laravel\\Forge\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Tests\\": "tests/"
         }
     },
     "extra": {
