@@ -1,33 +1,43 @@
 <?php
 
+namespace Tests;
 
-class ForgeSdkTest extends PHPUnit_Framework_TestCase
+use GuzzleHttp\Client;
+use GuzzleHttp\Psr7\Response;
+use Mockery;
+use PHPUnit\Framework\TestCase;
+use Laravel\Forge\Exceptions\FailedActionException;
+use Laravel\Forge\Exceptions\NotFoundException;
+use Laravel\Forge\Exceptions\ValidationException;
+use Laravel\Forge\Forge;
+
+class ForgeSDKTest extends TestCase
 {
-    public function tearDown()
+    protected function tearDown(): void
     {
         Mockery::close();
     }
 
     public function test_making_basic_requests()
     {
-        $forge = new \Themsaid\Forge\Forge('123', $http = Mockery::mock('GuzzleHttp\Client'));
+        $forge = new Forge('123', $http = Mockery::mock(Client::class));
 
         $http->shouldReceive('request')->once()->with('GET', 'recipes', [])->andReturn(
-            $response = Mockery::mock('GuzzleHttp\Psr7\Response')
+            $response = Mockery::mock(Response::class)
         );
 
         $response->shouldReceive('getStatusCode')->once()->andReturn(200);
         $response->shouldReceive('getBody')->once()->andReturn('{"recipes": [{"key": "value"}]}');
 
-        $forge->recipes();
+        $this->assertCount(1, $forge->recipes());
     }
 
     public function test_handling_validation_errors()
     {
-        $forge = new \Themsaid\Forge\Forge('123', $http = Mockery::mock('GuzzleHttp\Client'));
+        $forge = new Forge('123', $http = Mockery::mock(Client::class));
 
         $http->shouldReceive('request')->once()->with('GET', 'recipes', [])->andReturn(
-            $response = Mockery::mock('GuzzleHttp\Psr7\Response')
+            $response = Mockery::mock(Response::class)
         );
 
         $response->shouldReceive('getStatusCode')->andReturn(422);
@@ -35,7 +45,7 @@ class ForgeSdkTest extends PHPUnit_Framework_TestCase
 
         try {
             $forge->recipes();
-        } catch (\Themsaid\Forge\Exceptions\ValidationException $e) {
+        } catch (ValidationException $e) {
         }
 
         $this->assertEquals(['name' => ['The name is required.']], $e->errors());
@@ -43,12 +53,12 @@ class ForgeSdkTest extends PHPUnit_Framework_TestCase
 
     public function test_handling_404_errors()
     {
-        $this->expectException(\Themsaid\Forge\Exceptions\NotFoundException::class);
+        $this->expectException(NotFoundException::class);
 
-        $forge = new \Themsaid\Forge\Forge('123', $http = Mockery::mock('GuzzleHttp\Client'));
+        $forge = new Forge('123', $http = Mockery::mock(Client::class));
 
         $http->shouldReceive('request')->once()->with('GET', 'recipes', [])->andReturn(
-            $response = Mockery::mock('GuzzleHttp\Psr7\Response')
+            $response = Mockery::mock(Response::class)
         );
 
         $response->shouldReceive('getStatusCode')->andReturn(404);
@@ -58,10 +68,10 @@ class ForgeSdkTest extends PHPUnit_Framework_TestCase
 
     public function test_handling_failed_action_errors()
     {
-        $forge = new \Themsaid\Forge\Forge('123', $http = Mockery::mock('GuzzleHttp\Client'));
+        $forge = new Forge('123', $http = Mockery::mock(Client::class));
 
         $http->shouldReceive('request')->once()->with('GET', 'recipes', [])->andReturn(
-            $response = Mockery::mock('GuzzleHttp\Psr7\Response')
+            $response = Mockery::mock(Response::class)
         );
 
         $response->shouldReceive('getStatusCode')->andReturn(400);
@@ -69,9 +79,8 @@ class ForgeSdkTest extends PHPUnit_Framework_TestCase
 
         try {
             $forge->recipes();
-        } catch (\Themsaid\Forge\Exceptions\FailedActionException $e) {
+        } catch (FailedActionException $e) {
+            $this->assertEquals('Error!', $e->getMessage());
         }
-
-        $this->assertEquals('Error!', $e->getMessage());
     }
 }
