@@ -1,10 +1,10 @@
 <?php
 
-namespace Themsaid\Forge\Actions;
+namespace Laravel\Forge\Actions;
 
-use Themsaid\Forge\Resources\BackupConfiguration;
-use Themsaid\Forge\Resources\Event;
-use Themsaid\Forge\Resources\Server;
+use Laravel\Forge\Resources\BackupConfiguration;
+use Laravel\Forge\Resources\Event;
+use Laravel\Forge\Resources\Server;
 
 trait ManagesBackups
 {
